Initial import
At this point, the netlink module is okayish. When many routes are
removed/added at the same time, the socket receive buffer is too small
to handle that and the application panics. There are several
solutions, like increasing receive buffer size for the netlink
socket and/or (most likely and) handle the failure gracefully and poll
again the whole routing table.

However, wouldn't it be simple/more robust to just poll the routing
table for the prefixes we want? We cannot filter by prefix, so we'll
get the whole table at once. Another solution would be to rely on BIRD
and query BIRD for the prefix we want. This would be a major change
but it would fit our use case and it would be CPU-friendly.