netlink: handle buffer overflow problems
We use an FSM to restart listening socket if we get an overflow. We
also use a backoff algorithm to restart the whole process as well as a
cure timer to reset it.

Listening is done before requesting a dump. Therefore, we may have
some redundant events but the state should be eventually
consistent. We also add some buffer to channels to increase
performance.